Module aws_sdk_alexaforbusiness::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

ApproveSkill

Fluent builder constructing a request to ApproveSkill.

AssociateContactWithAddressBook

Fluent builder constructing a request to AssociateContactWithAddressBook.

AssociateDeviceWithNetworkProfile

Fluent builder constructing a request to AssociateDeviceWithNetworkProfile.

AssociateDeviceWithRoom

Fluent builder constructing a request to AssociateDeviceWithRoom.

AssociateSkillGroupWithRoom

Fluent builder constructing a request to AssociateSkillGroupWithRoom.

AssociateSkillWithSkillGroup

Fluent builder constructing a request to AssociateSkillWithSkillGroup.

AssociateSkillWithUsers

Fluent builder constructing a request to AssociateSkillWithUsers.

CreateAddressBook

Fluent builder constructing a request to CreateAddressBook.

CreateBusinessReportSchedule

Fluent builder constructing a request to CreateBusinessReportSchedule.

CreateConferenceProvider

Fluent builder constructing a request to CreateConferenceProvider.

CreateContact

Fluent builder constructing a request to CreateContact.

CreateGatewayGroup

Fluent builder constructing a request to CreateGatewayGroup.

CreateNetworkProfile

Fluent builder constructing a request to CreateNetworkProfile.

CreateProfile

Fluent builder constructing a request to CreateProfile.

CreateRoom

Fluent builder constructing a request to CreateRoom.

CreateSkillGroup

Fluent builder constructing a request to CreateSkillGroup.

CreateUser

Fluent builder constructing a request to CreateUser.

DeleteAddressBook

Fluent builder constructing a request to DeleteAddressBook.

DeleteBusinessReportSchedule

Fluent builder constructing a request to DeleteBusinessReportSchedule.

DeleteConferenceProvider

Fluent builder constructing a request to DeleteConferenceProvider.

DeleteContact

Fluent builder constructing a request to DeleteContact.

DeleteDevice

Fluent builder constructing a request to DeleteDevice.

DeleteDeviceUsageData

Fluent builder constructing a request to DeleteDeviceUsageData.

DeleteGatewayGroup

Fluent builder constructing a request to DeleteGatewayGroup.

DeleteNetworkProfile

Fluent builder constructing a request to DeleteNetworkProfile.

DeleteProfile

Fluent builder constructing a request to DeleteProfile.

DeleteRoom

Fluent builder constructing a request to DeleteRoom.

DeleteRoomSkillParameter

Fluent builder constructing a request to DeleteRoomSkillParameter.

DeleteSkillAuthorization

Fluent builder constructing a request to DeleteSkillAuthorization.

DeleteSkillGroup

Fluent builder constructing a request to DeleteSkillGroup.

DeleteUser

Fluent builder constructing a request to DeleteUser.

DisassociateContactFromAddressBook

Fluent builder constructing a request to DisassociateContactFromAddressBook.

DisassociateDeviceFromRoom

Fluent builder constructing a request to DisassociateDeviceFromRoom.

DisassociateSkillFromSkillGroup

Fluent builder constructing a request to DisassociateSkillFromSkillGroup.

DisassociateSkillFromUsers

Fluent builder constructing a request to DisassociateSkillFromUsers.

DisassociateSkillGroupFromRoom

Fluent builder constructing a request to DisassociateSkillGroupFromRoom.

ForgetSmartHomeAppliances

Fluent builder constructing a request to ForgetSmartHomeAppliances.

GetAddressBook

Fluent builder constructing a request to GetAddressBook.

GetConferencePreference

Fluent builder constructing a request to GetConferencePreference.

GetConferenceProvider

Fluent builder constructing a request to GetConferenceProvider.

GetContact

Fluent builder constructing a request to GetContact.

GetDevice

Fluent builder constructing a request to GetDevice.

GetGateway

Fluent builder constructing a request to GetGateway.

GetGatewayGroup

Fluent builder constructing a request to GetGatewayGroup.

GetInvitationConfiguration

Fluent builder constructing a request to GetInvitationConfiguration.

GetNetworkProfile

Fluent builder constructing a request to GetNetworkProfile.

GetProfile

Fluent builder constructing a request to GetProfile.

GetRoom

Fluent builder constructing a request to GetRoom.

GetRoomSkillParameter

Fluent builder constructing a request to GetRoomSkillParameter.

GetSkillGroup

Fluent builder constructing a request to GetSkillGroup.

ListBusinessReportSchedules

Fluent builder constructing a request to ListBusinessReportSchedules.

ListConferenceProviders

Fluent builder constructing a request to ListConferenceProviders.

ListDeviceEvents

Fluent builder constructing a request to ListDeviceEvents.

ListGatewayGroups

Fluent builder constructing a request to ListGatewayGroups.

ListGateways

Fluent builder constructing a request to ListGateways.

ListSkills

Fluent builder constructing a request to ListSkills.

ListSkillsStoreCategories

Fluent builder constructing a request to ListSkillsStoreCategories.

ListSkillsStoreSkillsByCategory

Fluent builder constructing a request to ListSkillsStoreSkillsByCategory.

ListSmartHomeAppliances

Fluent builder constructing a request to ListSmartHomeAppliances.

ListTags

Fluent builder constructing a request to ListTags.

PutConferencePreference

Fluent builder constructing a request to PutConferencePreference.

PutInvitationConfiguration

Fluent builder constructing a request to PutInvitationConfiguration.

PutRoomSkillParameter

Fluent builder constructing a request to PutRoomSkillParameter.

PutSkillAuthorization

Fluent builder constructing a request to PutSkillAuthorization.

RegisterAVSDevice

Fluent builder constructing a request to RegisterAVSDevice.

RejectSkill

Fluent builder constructing a request to RejectSkill.

ResolveRoom

Fluent builder constructing a request to ResolveRoom.

RevokeInvitation

Fluent builder constructing a request to RevokeInvitation.

SearchAddressBooks

Fluent builder constructing a request to SearchAddressBooks.

SearchContacts

Fluent builder constructing a request to SearchContacts.

SearchDevices

Fluent builder constructing a request to SearchDevices.

SearchNetworkProfiles

Fluent builder constructing a request to SearchNetworkProfiles.

SearchProfiles

Fluent builder constructing a request to SearchProfiles.

SearchRooms

Fluent builder constructing a request to SearchRooms.

SearchSkillGroups

Fluent builder constructing a request to SearchSkillGroups.

SearchUsers

Fluent builder constructing a request to SearchUsers.

SendAnnouncement

Fluent builder constructing a request to SendAnnouncement.

SendInvitation

Fluent builder constructing a request to SendInvitation.

StartDeviceSync

Fluent builder constructing a request to StartDeviceSync.

StartSmartHomeApplianceDiscovery

Fluent builder constructing a request to StartSmartHomeApplianceDiscovery.

TagResource

Fluent builder constructing a request to TagResource.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateAddressBook

Fluent builder constructing a request to UpdateAddressBook.

UpdateBusinessReportSchedule

Fluent builder constructing a request to UpdateBusinessReportSchedule.

UpdateConferenceProvider

Fluent builder constructing a request to UpdateConferenceProvider.

UpdateContact

Fluent builder constructing a request to UpdateContact.

UpdateDevice

Fluent builder constructing a request to UpdateDevice.

UpdateGateway

Fluent builder constructing a request to UpdateGateway.

UpdateGatewayGroup

Fluent builder constructing a request to UpdateGatewayGroup.

UpdateNetworkProfile

Fluent builder constructing a request to UpdateNetworkProfile.

UpdateProfile

Fluent builder constructing a request to UpdateProfile.

UpdateRoom

Fluent builder constructing a request to UpdateRoom.

UpdateSkillGroup

Fluent builder constructing a request to UpdateSkillGroup.