Struct aws_sdk_alexaforbusiness::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AlexaForBusiness.

This client allows ergonomic access to a AlexaForBusiness-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn approve_skill(&self) -> ApproveSkill<C, M, R>

pub fn associate_contact_with_address_book(
&self
) -> AssociateContactWithAddressBook<C, M, R>

pub fn associate_device_with_network_profile(
&self
) -> AssociateDeviceWithNetworkProfile<C, M, R>

pub fn associate_device_with_room(&self) -> AssociateDeviceWithRoom<C, M, R>

pub fn associate_skill_group_with_room(
&self
) -> AssociateSkillGroupWithRoom<C, M, R>

pub fn associate_skill_with_skill_group(
&self
) -> AssociateSkillWithSkillGroup<C, M, R>

pub fn associate_skill_with_users(&self) -> AssociateSkillWithUsers<C, M, R>

pub fn create_address_book(&self) -> CreateAddressBook<C, M, R>

pub fn create_business_report_schedule(
&self
) -> CreateBusinessReportSchedule<C, M, R>

pub fn create_conference_provider(&self) -> CreateConferenceProvider<C, M, R>

pub fn create_contact(&self) -> CreateContact<C, M, R>

pub fn create_gateway_group(&self) -> CreateGatewayGroup<C, M, R>

pub fn create_network_profile(&self) -> CreateNetworkProfile<C, M, R>

pub fn create_profile(&self) -> CreateProfile<C, M, R>

pub fn create_room(&self) -> CreateRoom<C, M, R>

pub fn create_skill_group(&self) -> CreateSkillGroup<C, M, R>

pub fn create_user(&self) -> CreateUser<C, M, R>

pub fn delete_address_book(&self) -> DeleteAddressBook<C, M, R>

pub fn delete_business_report_schedule(
&self
) -> DeleteBusinessReportSchedule<C, M, R>

pub fn delete_conference_provider(&self) -> DeleteConferenceProvider<C, M, R>

pub fn delete_contact(&self) -> DeleteContact<C, M, R>

pub fn delete_device(&self) -> DeleteDevice<C, M, R>

pub fn delete_device_usage_data(&self) -> DeleteDeviceUsageData<C, M, R>

pub fn delete_gateway_group(&self) -> DeleteGatewayGroup<C, M, R>

pub fn delete_network_profile(&self) -> DeleteNetworkProfile<C, M, R>

pub fn delete_profile(&self) -> DeleteProfile<C, M, R>

pub fn delete_room(&self) -> DeleteRoom<C, M, R>

pub fn delete_room_skill_parameter(&self) -> DeleteRoomSkillParameter<C, M, R>

pub fn delete_skill_authorization(&self) -> DeleteSkillAuthorization<C, M, R>

pub fn delete_skill_group(&self) -> DeleteSkillGroup<C, M, R>

pub fn delete_user(&self) -> DeleteUser<C, M, R>

pub fn disassociate_contact_from_address_book(
&self
) -> DisassociateContactFromAddressBook<C, M, R>

pub fn disassociate_device_from_room(
&self
) -> DisassociateDeviceFromRoom<C, M, R>

pub fn disassociate_skill_from_skill_group(
&self
) -> DisassociateSkillFromSkillGroup<C, M, R>

pub fn disassociate_skill_from_users(
&self
) -> DisassociateSkillFromUsers<C, M, R>

pub fn disassociate_skill_group_from_room(
&self
) -> DisassociateSkillGroupFromRoom<C, M, R>

pub fn forget_smart_home_appliances(&self) -> ForgetSmartHomeAppliances<C, M, R>

pub fn get_address_book(&self) -> GetAddressBook<C, M, R>

pub fn get_conference_preference(&self) -> GetConferencePreference<C, M, R>

pub fn get_conference_provider(&self) -> GetConferenceProvider<C, M, R>

pub fn get_contact(&self) -> GetContact<C, M, R>

pub fn get_device(&self) -> GetDevice<C, M, R>

pub fn get_gateway(&self) -> GetGateway<C, M, R>

pub fn get_gateway_group(&self) -> GetGatewayGroup<C, M, R>

pub fn get_invitation_configuration(
&self
) -> GetInvitationConfiguration<C, M, R>

pub fn get_network_profile(&self) -> GetNetworkProfile<C, M, R>

pub fn get_profile(&self) -> GetProfile<C, M, R>

pub fn get_room(&self) -> GetRoom<C, M, R>

pub fn get_room_skill_parameter(&self) -> GetRoomSkillParameter<C, M, R>

pub fn get_skill_group(&self) -> GetSkillGroup<C, M, R>

pub fn list_business_report_schedules(
&self
) -> ListBusinessReportSchedules<C, M, R>

pub fn list_conference_providers(&self) -> ListConferenceProviders<C, M, R>

pub fn list_device_events(&self) -> ListDeviceEvents<C, M, R>

pub fn list_gateway_groups(&self) -> ListGatewayGroups<C, M, R>

pub fn list_gateways(&self) -> ListGateways<C, M, R>

pub fn list_skills(&self) -> ListSkills<C, M, R>

pub fn list_skills_store_categories(&self) -> ListSkillsStoreCategories<C, M, R>

pub fn list_skills_store_skills_by_category(
&self
) -> ListSkillsStoreSkillsByCategory<C, M, R>

pub fn list_smart_home_appliances(&self) -> ListSmartHomeAppliances<C, M, R>

pub fn list_tags(&self) -> ListTags<C, M, R>

pub fn put_conference_preference(&self) -> PutConferencePreference<C, M, R>

pub fn put_invitation_configuration(
&self
) -> PutInvitationConfiguration<C, M, R>

pub fn put_room_skill_parameter(&self) -> PutRoomSkillParameter<C, M, R>

pub fn put_skill_authorization(&self) -> PutSkillAuthorization<C, M, R>

pub fn register_avs_device(&self) -> RegisterAVSDevice<C, M, R>

pub fn reject_skill(&self) -> RejectSkill<C, M, R>

pub fn resolve_room(&self) -> ResolveRoom<C, M, R>

pub fn revoke_invitation(&self) -> RevokeInvitation<C, M, R>

pub fn search_address_books(&self) -> SearchAddressBooks<C, M, R>

pub fn search_contacts(&self) -> SearchContacts<C, M, R>

pub fn search_devices(&self) -> SearchDevices<C, M, R>

pub fn search_network_profiles(&self) -> SearchNetworkProfiles<C, M, R>

pub fn search_profiles(&self) -> SearchProfiles<C, M, R>

pub fn search_rooms(&self) -> SearchRooms<C, M, R>

pub fn search_skill_groups(&self) -> SearchSkillGroups<C, M, R>

pub fn search_users(&self) -> SearchUsers<C, M, R>

pub fn send_announcement(&self) -> SendAnnouncement<C, M, R>

pub fn send_invitation(&self) -> SendInvitation<C, M, R>

pub fn start_device_sync(&self) -> StartDeviceSync<C, M, R>

pub fn start_smart_home_appliance_discovery(
&self
) -> StartSmartHomeApplianceDiscovery<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_address_book(&self) -> UpdateAddressBook<C, M, R>

pub fn update_business_report_schedule(
&self
) -> UpdateBusinessReportSchedule<C, M, R>

pub fn update_conference_provider(&self) -> UpdateConferenceProvider<C, M, R>

pub fn update_contact(&self) -> UpdateContact<C, M, R>

pub fn update_device(&self) -> UpdateDevice<C, M, R>

pub fn update_gateway(&self) -> UpdateGateway<C, M, R>

pub fn update_gateway_group(&self) -> UpdateGatewayGroup<C, M, R>

pub fn update_network_profile(&self) -> UpdateNetworkProfile<C, M, R>

pub fn update_profile(&self) -> UpdateProfile<C, M, R>

pub fn update_room(&self) -> UpdateRoom<C, M, R>

pub fn update_skill_group(&self) -> UpdateSkillGroup<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self