Module aws_sdk_alexaforbusiness::error

Modules

already_exists_exception

See AlreadyExistsException

concurrent_modification_exception

See ConcurrentModificationException

device_not_registered_exception

See DeviceNotRegisteredException

invalid_certificate_authority_exception

See InvalidCertificateAuthorityException

invalid_device_exception

See InvalidDeviceException

invalid_secrets_manager_resource_exception

See InvalidSecretsManagerResourceException

invalid_service_linked_role_state_exception

See InvalidServiceLinkedRoleStateException

invalid_user_status_exception

See InvalidUserStatusException

limit_exceeded_exception

See LimitExceededException

name_in_use_exception

See NameInUseException

not_found_exception

See NotFoundException

resource_associated_exception

See ResourceAssociatedException

resource_in_use_exception

See ResourceInUseException

skill_not_linked_exception

See SkillNotLinkedException

unauthorized_exception

See UnauthorizedException

Structs

AlreadyExistsException

The resource being created already exists.

ApproveSkillError

AssociateContactWithAddressBookError

AssociateDeviceWithNetworkProfileError

AssociateDeviceWithRoomError

AssociateSkillGroupWithRoomError

AssociateSkillWithSkillGroupError

AssociateSkillWithUsersError

ConcurrentModificationException

There is a concurrent modification of resources.

CreateAddressBookError

CreateBusinessReportScheduleError

CreateConferenceProviderError

CreateContactError

CreateGatewayGroupError

CreateNetworkProfileError

CreateProfileError

CreateRoomError

CreateSkillGroupError

CreateUserError

DeleteAddressBookError

DeleteBusinessReportScheduleError

DeleteConferenceProviderError

DeleteContactError

DeleteDeviceError

DeleteDeviceUsageDataError

DeleteGatewayGroupError

DeleteNetworkProfileError

DeleteProfileError

DeleteRoomError

DeleteRoomSkillParameterError

DeleteSkillAuthorizationError

DeleteSkillGroupError

DeleteUserError

DeviceNotRegisteredException

The request failed because this device is no longer registered and therefore no longer managed by this account.

DisassociateContactFromAddressBookError

DisassociateDeviceFromRoomError

DisassociateSkillFromSkillGroupError

DisassociateSkillFromUsersError

DisassociateSkillGroupFromRoomError

ForgetSmartHomeAppliancesError

GetAddressBookError

GetConferencePreferenceError

GetConferenceProviderError

GetContactError

GetDeviceError

GetGatewayError

GetGatewayGroupError

GetInvitationConfigurationError

GetNetworkProfileError

GetProfileError

GetRoomError

GetRoomSkillParameterError

GetSkillGroupError

InvalidCertificateAuthorityException

The Certificate Authority can't issue or revoke a certificate.

InvalidDeviceException

The device is in an invalid state.

InvalidSecretsManagerResourceException

A password in SecretsManager is in an invalid state.

InvalidServiceLinkedRoleStateException

The service linked role is locked for deletion.

InvalidUserStatusException

The attempt to update a user is invalid due to the user's current status.

LimitExceededException

You are performing an action that would put you beyond your account's limits.

ListBusinessReportSchedulesError

ListConferenceProvidersError

ListDeviceEventsError

ListGatewayGroupsError

ListGatewaysError

ListSkillsError

ListSkillsStoreCategoriesError

ListSkillsStoreSkillsByCategoryError

ListSmartHomeAppliancesError

ListTagsError

NameInUseException

The name sent in the request is already in use.

NotFoundException

The resource is not found.

PutConferencePreferenceError

PutInvitationConfigurationError

PutRoomSkillParameterError

PutSkillAuthorizationError

RegisterAVSDeviceError

RejectSkillError

ResolveRoomError

ResourceAssociatedException

Another resource is associated with the resource in the request.

ResourceInUseException

The resource in the request is already in use.

RevokeInvitationError

SearchAddressBooksError

SearchContactsError

SearchDevicesError

SearchNetworkProfilesError

SearchProfilesError

SearchRoomsError

SearchSkillGroupsError

SearchUsersError

SendAnnouncementError

SendInvitationError

SkillNotLinkedException

The skill must be linked to a third-party account.

StartDeviceSyncError

StartSmartHomeApplianceDiscoveryError

TagResourceError

UnauthorizedException

The caller has no permissions to operate on the resource involved in the API call.

UntagResourceError

UpdateAddressBookError

UpdateBusinessReportScheduleError

UpdateConferenceProviderError

UpdateContactError

UpdateDeviceError

UpdateGatewayError

UpdateGatewayGroupError

UpdateNetworkProfileError

UpdateProfileError

UpdateRoomError

UpdateSkillGroupError

Enums

ApproveSkillErrorKind

AssociateContactWithAddressBookErrorKind

AssociateDeviceWithNetworkProfileErrorKind

AssociateDeviceWithRoomErrorKind

AssociateSkillGroupWithRoomErrorKind

AssociateSkillWithSkillGroupErrorKind

AssociateSkillWithUsersErrorKind

CreateAddressBookErrorKind

CreateBusinessReportScheduleErrorKind

CreateConferenceProviderErrorKind

CreateContactErrorKind

CreateGatewayGroupErrorKind

CreateNetworkProfileErrorKind

CreateProfileErrorKind

CreateRoomErrorKind

CreateSkillGroupErrorKind

CreateUserErrorKind

DeleteAddressBookErrorKind

DeleteBusinessReportScheduleErrorKind

DeleteConferenceProviderErrorKind

DeleteContactErrorKind

DeleteDeviceErrorKind

DeleteDeviceUsageDataErrorKind

DeleteGatewayGroupErrorKind

DeleteNetworkProfileErrorKind

DeleteProfileErrorKind

DeleteRoomErrorKind

DeleteRoomSkillParameterErrorKind

DeleteSkillAuthorizationErrorKind

DeleteSkillGroupErrorKind

DeleteUserErrorKind

DisassociateContactFromAddressBookErrorKind

DisassociateDeviceFromRoomErrorKind

DisassociateSkillFromSkillGroupErrorKind

DisassociateSkillFromUsersErrorKind

DisassociateSkillGroupFromRoomErrorKind

ForgetSmartHomeAppliancesErrorKind

GetAddressBookErrorKind

GetConferencePreferenceErrorKind

GetConferenceProviderErrorKind

GetContactErrorKind

GetDeviceErrorKind

GetGatewayErrorKind

GetGatewayGroupErrorKind

GetInvitationConfigurationErrorKind

GetNetworkProfileErrorKind

GetProfileErrorKind

GetRoomErrorKind

GetRoomSkillParameterErrorKind

GetSkillGroupErrorKind

ListBusinessReportSchedulesErrorKind

ListConferenceProvidersErrorKind

ListDeviceEventsErrorKind

ListGatewayGroupsErrorKind

ListGatewaysErrorKind

ListSkillsErrorKind

ListSkillsStoreCategoriesErrorKind

ListSkillsStoreSkillsByCategoryErrorKind

ListSmartHomeAppliancesErrorKind

ListTagsErrorKind

PutConferencePreferenceErrorKind

PutInvitationConfigurationErrorKind

PutRoomSkillParameterErrorKind

PutSkillAuthorizationErrorKind

RegisterAVSDeviceErrorKind

RejectSkillErrorKind

ResolveRoomErrorKind

RevokeInvitationErrorKind

SearchAddressBooksErrorKind

SearchContactsErrorKind

SearchDevicesErrorKind

SearchNetworkProfilesErrorKind

SearchProfilesErrorKind

SearchRoomsErrorKind

SearchSkillGroupsErrorKind

SearchUsersErrorKind

SendAnnouncementErrorKind

SendInvitationErrorKind

StartDeviceSyncErrorKind

StartSmartHomeApplianceDiscoveryErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateAddressBookErrorKind

UpdateBusinessReportScheduleErrorKind

UpdateConferenceProviderErrorKind

UpdateContactErrorKind

UpdateDeviceErrorKind

UpdateGatewayErrorKind

UpdateGatewayGroupErrorKind

UpdateNetworkProfileErrorKind

UpdateProfileErrorKind

UpdateRoomErrorKind

UpdateSkillGroupErrorKind