Module aws_sdk_alexaforbusiness::model

Modules

address_book

See AddressBook

address_book_data

See AddressBookData

audio

See Audio

business_report

See BusinessReport

business_report_content_range

See BusinessReportContentRange

business_report_recurrence

See BusinessReportRecurrence

business_report_s3_location

See BusinessReportS3Location

business_report_schedule

See BusinessReportSchedule

category

See Category

conference_preference

See ConferencePreference

conference_provider

See ConferenceProvider

contact

See Contact

contact_data

See ContactData

content

See Content

create_end_of_meeting_reminder

See CreateEndOfMeetingReminder

create_instant_booking

See CreateInstantBooking

create_meeting_room_configuration

See CreateMeetingRoomConfiguration

create_require_check_in

See CreateRequireCheckIn

developer_info

See DeveloperInfo

device

See Device

device_data

See DeviceData

device_event

See DeviceEvent

device_network_profile_info

See DeviceNetworkProfileInfo

device_status_detail

See DeviceStatusDetail

device_status_info

See DeviceStatusInfo

end_of_meeting_reminder

See EndOfMeetingReminder

filter

See Filter

gateway

See Gateway

gateway_group

See GatewayGroup

gateway_group_summary

See GatewayGroupSummary

gateway_summary

See GatewaySummary

instant_booking

See InstantBooking

ip_dial_in

See IpDialIn

meeting_room_configuration

See MeetingRoomConfiguration

meeting_setting

See MeetingSetting

network_profile

See NetworkProfile

network_profile_data

See NetworkProfileData

phone_number

See PhoneNumber

profile

See Profile

profile_data

See ProfileData

pstn_dial_in

See PstnDialIn

require_check_in

See RequireCheckIn

room

See Room

room_data

See RoomData

room_skill_parameter

See RoomSkillParameter

sip_address

See SipAddress

skill_details

See SkillDetails

skill_group

See SkillGroup

skill_group_data

See SkillGroupData

skill_summary

See SkillSummary

skills_store_skill

See SkillsStoreSkill

smart_home_appliance

See SmartHomeAppliance

sort

See Sort

ssml

See Ssml

tag

See Tag

text

See Text

update_end_of_meeting_reminder

See UpdateEndOfMeetingReminder

update_instant_booking

See UpdateInstantBooking

update_meeting_room_configuration

See UpdateMeetingRoomConfiguration

update_require_check_in

See UpdateRequireCheckIn

user_data

See UserData

Structs

AddressBook

An address book with attributes.

AddressBookData

Information related to an address book.

Audio

The audio message. There is a 1 MB limit on the audio file input and the only supported

BusinessReport

Usage report with specified parameters.

BusinessReportContentRange

The content range of the report.

BusinessReportRecurrence

The recurrence of the reports.

BusinessReportS3Location

The S3 location of the output reports.

BusinessReportSchedule

The schedule of the usage report.

Category

The skill store category that is shown. Alexa skills are assigned a specific skill

ConferencePreference

The default conference provider that is used if no other scheduled meetings are

ConferenceProvider

An entity that provides a conferencing solution. Alexa for Business acts as the voice interface and mediator that connects users to their preferred conference provider. Examples of conference providers include Amazon Chime, Zoom, Cisco, and Polycom.

Contact

A contact with attributes.

ContactData

Information related to a contact.

Content

The content definition. This can contain only one text, SSML, or audio list

CreateEndOfMeetingReminder

Creates settings for the end of meeting reminder feature that are applied to a room

CreateInstantBooking

Creates settings for the instant booking feature that are applied to a room profile.

CreateMeetingRoomConfiguration

Creates meeting room settings of a room profile.

CreateRequireCheckIn

Creates settings for the require check in feature that are applied to a room profile. Require check in allows a meeting room’s Alexa or AVS device to prompt the user to check in; otherwise, the room will be released.

DeveloperInfo

The details about the developer that published the skill.

Device

A device with attributes.

DeviceData

Device attributes.

DeviceEvent

The list of device events.

DeviceNetworkProfileInfo

Detailed information about a device's network profile.

DeviceStatusDetail

Details of a device’s status.

DeviceStatusInfo

Detailed information about a device's status.

EndOfMeetingReminder

Settings for the end of meeting reminder feature that are applied to a room profile. The

Filter

A filter name and value pair that is used to return a more specific list of results.

Gateway

The details of the gateway.

GatewayGroup

The details of the gateway group.

GatewayGroupSummary

The summary of a gateway group.

GatewaySummary

The summary of a gateway.

InstantBooking

Settings for the instant booking feature that are applied to a room profile. When users

IpDialIn

The IP endpoint and protocol for calling.

MeetingRoomConfiguration

Meeting room settings of a room profile.

MeetingSetting

The values that indicate whether a pin is always required (YES), never required (NO),

NetworkProfile

The network profile associated with a device.

NetworkProfileData

The data associated with a network profile.

PhoneNumber

The phone number for the contact containing the raw number and phone number type.

Profile

A room profile with attributes.

ProfileData

The data of a room profile.

PstnDialIn

The information for public switched telephone network (PSTN) conferencing.

RequireCheckIn

Settings for the require check in feature that are applied to a room profile. Require check in allows a meeting room’s Alexa or AVS device to prompt the user to check in; otherwise, the room will be released.

Room

A room with attributes.

RoomData

The data of a room.

RoomSkillParameter

A skill parameter associated with a room.

SipAddress

The SIP address for the contact containing the URI and SIP address type.

SkillDetails

Granular information about the skill.

SkillGroup

A skill group with attributes.

SkillGroupData

The attributes of a skill group.

SkillSummary

The summary of skills.

SkillsStoreSkill

The detailed information about an Alexa skill.

SmartHomeAppliance

A smart home appliance that can connect to a central system. Any domestic device can be

Sort

An object representing a sort criteria.

Ssml

The SSML message. For more information, see SSML Reference.

Tag

A key-value pair that can be associated with a resource.

Text

The text message.

UpdateEndOfMeetingReminder

Settings for the end of meeting reminder feature that are applied to a room profile. The

UpdateInstantBooking

Updates settings for the instant booking feature that are applied to a room profile. If

UpdateMeetingRoomConfiguration

Updates meeting room settings of a room profile.

UpdateRequireCheckIn

Updates settings for the require check in feature that are applied to a room profile. Require check in allows a meeting room’s Alexa or AVS device to prompt the user to check in; otherwise, the room will be released.

UserData

Information related to a user.

Enums

BusinessReportFailureCode

BusinessReportFormat

BusinessReportInterval

BusinessReportStatus

CommsProtocol

ConferenceProviderType

ConnectionStatus

DeviceEventType

DeviceStatus

DeviceStatusDetailCode

DeviceUsageType

DistanceUnit

EnablementType

EnablementTypeFilter

EndOfMeetingReminderType

EnrollmentStatus

Feature

Locale

NetworkEapMethod

NetworkSecurityType

PhoneNumberType

RequirePin

SipType

SkillType

SkillTypeFilter

SortValue

TemperatureUnit

WakeWord