Module aws_sdk_alexaforbusiness::operation

Structs

ApproveSkill

Associates a skill with the organization under the customer's AWS account. If a skill

AssociateContactWithAddressBook

Associates a contact with a given address book.

AssociateDeviceWithNetworkProfile

Associates a device with the specified network profile.

AssociateDeviceWithRoom

Associates a device with a given room. This applies all the settings from the room

AssociateSkillGroupWithRoom

Associates a skill group with a given room. This enables all skills in the associated

AssociateSkillWithSkillGroup

Associates a skill with a skill group.

AssociateSkillWithUsers

Makes a private skill available for enrolled users to enable on their devices.

CreateAddressBook

Creates an address book with the specified details.

CreateBusinessReportSchedule

Creates a recurring schedule for usage reports to deliver to the specified S3

CreateConferenceProvider

Adds a new conference provider under the user's AWS account.

CreateContact

Creates a contact with the specified details.

CreateGatewayGroup

Creates a gateway group with the specified details.

CreateNetworkProfile

Creates a network profile with the specified details.

CreateProfile

Creates a new room profile with the specified details.

CreateRoom

Creates a room with the specified details.

CreateSkillGroup

Creates a skill group with a specified name and description.

CreateUser

Creates a user.

DeleteAddressBook

Deletes an address book by the address book ARN.

DeleteBusinessReportSchedule

Deletes the recurring report delivery schedule with the specified schedule

DeleteConferenceProvider

Deletes a conference provider.

DeleteContact

Deletes a contact by the contact ARN.

DeleteDevice

Removes a device from Alexa For Business.

DeleteDeviceUsageData

When this action is called for a specified shared device, it allows authorized users to

DeleteGatewayGroup

Deletes a gateway group.

DeleteNetworkProfile

Deletes a network profile by the network profile ARN.

DeleteProfile

Deletes a room profile by the profile ARN.

DeleteRoom

Deletes a room by the room ARN.

DeleteRoomSkillParameter

Deletes room skill parameter details by room, skill, and parameter key ID.

DeleteSkillAuthorization

Unlinks a third-party account from a skill.

DeleteSkillGroup

Deletes a skill group by skill group ARN.

DeleteUser

Deletes a specified user by user ARN and enrollment ARN.

DisassociateContactFromAddressBook

Disassociates a contact from a given address book.

DisassociateDeviceFromRoom

Disassociates a device from its current room. The device continues to be connected to

DisassociateSkillFromSkillGroup

Disassociates a skill from a skill group.

DisassociateSkillFromUsers

Makes a private skill unavailable for enrolled users and prevents them from enabling it

DisassociateSkillGroupFromRoom

Disassociates a skill group from a specified room. This disables all skills in the

ForgetSmartHomeAppliances

Forgets smart home appliances associated to a room.

GetAddressBook

Gets address the book details by the address book ARN.

GetConferencePreference

Retrieves the existing conference preferences.

GetConferenceProvider

Gets details about a specific conference provider.

GetContact

Gets the contact details by the contact ARN.

GetDevice

Gets the details of a device by device ARN.

GetGateway

Retrieves the details of a gateway.

GetGatewayGroup

Retrieves the details of a gateway group.

GetInvitationConfiguration

Retrieves the configured values for the user enrollment invitation email

GetNetworkProfile

Gets the network profile details by the network profile ARN.

GetProfile

Gets the details of a room profile by profile ARN.

GetRoom

Gets room details by room ARN.

GetRoomSkillParameter

Gets room skill parameter details by room, skill, and parameter key ARN.

GetSkillGroup

Gets skill group details by skill group ARN.

ListBusinessReportSchedules

Lists the details of the schedules that a user configured. A download URL of the report associated with each schedule is returned every time this action is called. A new download URL is returned each time, and is valid for 24 hours.

ListConferenceProviders

Lists conference providers under a specific AWS account.

ListDeviceEvents

Lists the device event history, including device connection status, for up to 30

ListGatewayGroups

Retrieves a list of gateway group summaries. Use GetGatewayGroup to retrieve details of

ListGateways

Retrieves a list of gateway summaries. Use GetGateway to retrieve details of a specific

ListSkills

Lists all enabled skills in a specific skill group.

ListSkillsStoreCategories

Lists all categories in the Alexa skill store.

ListSkillsStoreSkillsByCategory

Lists all skills in the Alexa skill store by category.

ListSmartHomeAppliances

Lists all of the smart home appliances associated with a room.

ListTags

Lists all tags for the specified resource.

PutConferencePreference

Sets the conference preferences on a specific conference provider at the account

PutInvitationConfiguration

Configures the email template for the user enrollment invitation with the specified

PutRoomSkillParameter

Updates room skill parameter details by room, skill, and parameter key ID. Not all

PutSkillAuthorization

Links a user's account to a third-party skill provider. If this API operation is

RegisterAVSDevice

Registers an Alexa-enabled device built by an Original Equipment Manufacturer (OEM)

RejectSkill

Disassociates a skill from the organization under a user's AWS account. If the skill

ResolveRoom

Determines the details for the room from which a skill request was invoked. This

RevokeInvitation

Revokes an invitation and invalidates the enrollment URL.

SearchAddressBooks

Searches address books and lists the ones that meet a set of filter and sort

SearchContacts

Searches contacts and lists the ones that meet a set of filter and sort

SearchDevices

Searches devices and lists the ones that meet a set of filter criteria.

SearchNetworkProfiles

Searches network profiles and lists the ones that meet a set of filter and sort

SearchProfiles

Searches room profiles and lists the ones that meet a set of filter

SearchRooms

Searches rooms and lists the ones that meet a set of filter and sort

SearchSkillGroups

Searches skill groups and lists the ones that meet a set of filter and sort

SearchUsers

Searches users and lists the ones that meet a set of filter and sort

SendAnnouncement

Triggers an asynchronous flow to send text, SSML, or audio announcements to rooms that

SendInvitation

Sends an enrollment invitation email with a URL to a user. The URL is valid for 30

StartDeviceSync

Resets a device and its account to the known default settings. This clears all

StartSmartHomeApplianceDiscovery

Initiates the discovery of any smart home appliances associated with the

TagResource

Adds metadata tags to a specified resource.

UntagResource

Removes metadata tags from a specified resource.

UpdateAddressBook

Updates address book details by the address book ARN.

UpdateBusinessReportSchedule

Updates the configuration of the report delivery schedule with the specified schedule

UpdateConferenceProvider

Updates an existing conference provider's settings.

UpdateContact

Updates the contact details by the contact ARN.

UpdateDevice

Updates the device name by device ARN.

UpdateGateway

Updates the details of a gateway. If any optional field is not provided, the existing

UpdateGatewayGroup

Updates the details of a gateway group. If any optional field is not provided, the

UpdateNetworkProfile

Updates a network profile by the network profile ARN.

UpdateProfile

Updates an existing room profile by room profile ARN.

UpdateRoom

Updates room details by room ARN.

UpdateSkillGroup

Updates skill group details by skill group ARN.